+++ js/main.js
@@ -1,5 +1,7 @@
 import { SIMILAR_OBJECTS } from './data.js';
 import { checkStringLength } from './utils.js';
+import { renderingPhoto } from './rendering-photo.js';
 
 SIMILAR_OBJECTS(25);
 checkStringLength('Hello world', 10);
+renderingPhoto();
